requirements.txt
absl-py==2.0.0
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cycler==0.11.0
dlib==19.24.2
exceptiongroup==1.1.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.103.0
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.42.1
gast==0.5.4
google-auth==2.23.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
Jinja2==3.1.2
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
mediapipe==0.10.5
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.1
numpy==1.25.2
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
Pillow==10.0.1
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
starlette==0.27.0
sympy==1.12
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
torch==2.0.1
typing_extensions==4.7.1
urllib3==1.26.6
uvicorn==0.23.2
Werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0
